Fix search for props in right direction
Say the user is editing a placeholder causing us to have to update a
mirror. From parsing the snippet we know if the mirror is above or below
the placeholder, and searching from the placeholder in the right
direction can be much better than, say, searching downward from the
first line of the file.