Add backward jumping
Currently works by popping snippet instances until a placeholder to jump
to is found. For some use cases it will be better not to pop the stack
when jumping backward, and vice versa. Always popping leads to simpler
implementation, but we will have see how it turns out.

For the record, UltiSnips does not allow backward jumping out of
snippets.